Update Command class and tests, pull class into separate file
